requirements.txt
Django
docutils
python-memcached
Sphinx
django-markup-deprecated
Pillow  # Required to generate sprites from Graphite
